1. Parsing JSON data from a RESTful Web Service. DONE. 2. Download file. DONE
Next steps: we need to use the SQLITE database. There, we are going to store
the index for every downloaded file. With the advertisement id retrieved from
the JSON RESTful response, we have a unique identifier for the downloaded ad.
We are going to use that id to find out if we have to download a file because
we do not have it on the mobile or we can skip the download because it was already
downloaded.